refactor(storefront/product): move product's variant selection hook
